webhook-listener/requirements.txt
flask==3.0
gunicorn==21
pyyaml==5.1
